feat: fix options support
